+++ server/src/main/java/org/opensearch/common/settings/ClusterSettings.java
@@ -129,6 +129,7 @@ import org.opensearch.monitor.process.ProcessService;
 import org.opensearch.node.Node;
 import org.opensearch.node.Node.DiscoverySettings;
 import org.opensearch.node.NodeRoleSettings;
+import org.opensearch.node.remotestore.RemoteStoreNodeService;
 import org.opensearch.persistent.PersistentTasksClusterService;
 import org.opensearch.persistent.decider.EnableAssignmentDecider;
 import org.opensearch.plugins.PluginsService;
@@ -670,7 +671,9 @@ public final class ClusterSettings extends AbstractScopedSettings {
                 // Remote cluster state settings
                 RemoteClusterStateService.REMOTE_CLUSTER_STATE_ENABLED_SETTING,
                 RemoteClusterStateService.REMOTE_CLUSTER_STATE_REPOSITORY_SETTING,
-
+              
+                RemoteStoreNodeService.REMOTE_STORE_COMPATIBILITY_MODE_SETTING
+              
                 // Remote store enablement settings
                 IndicesService.CLUSTER_REMOTE_STORE_ENABLED_SETTING,
                 IndicesService.CLUSTER_REMOTE_SEGMENT_STORE_REPOSITORY_SETTING,