+++ server/src/main/java/org/opensearch/index/remote/RemoteSegmentStats.java
@@ -256,7 +256,7 @@ public class RemoteSegmentStats implements Writeable, ToXContentFragment {
     }
 
     private void buildUploadStats(XContentBuilder builder) throws IOException {
-        builder.startObject(Fields.TOTAL_UPLOADS);
+        builder.startObject(Fields.TOTAL_UPLOAD_SIZE);
         builder.humanReadableField(Fields.STARTED_BYTES, Fields.STARTED, new ByteSizeValue(uploadBytesStarted));
         builder.humanReadableField(Fields.SUCCEEDED_BYTES, Fields.SUCCEEDED, new ByteSizeValue(uploadBytesSucceeded));
         builder.humanReadableField(Fields.FAILED_BYTES, Fields.FAILED, new ByteSizeValue(uploadBytesFailed));
@@ -270,7 +270,7 @@ public class RemoteSegmentStats implements Writeable, ToXContentFragment {
     }
 
     private void buildDownloadStats(XContentBuilder builder) throws IOException {
-        builder.startObject(Fields.TOTAL_DOWNLOADS);
+        builder.startObject(Fields.TOTAL_DOWNLOAD_SIZE);
         builder.humanReadableField(Fields.STARTED_BYTES, Fields.STARTED, new ByteSizeValue(downloadBytesStarted));
         builder.humanReadableField(Fields.SUCCEEDED_BYTES, Fields.SUCCEEDED, new ByteSizeValue(downloadBytesSucceeded));
         builder.humanReadableField(Fields.FAILED_BYTES, Fields.FAILED, new ByteSizeValue(downloadBytesFailed));
@@ -282,8 +282,8 @@ public class RemoteSegmentStats implements Writeable, ToXContentFragment {
         static final String REMOTE_STORE = "remote_store";
         static final String UPLOAD = "upload";
         static final String DOWNLOAD = "download";
-        static final String TOTAL_UPLOADS = "total_uploads";
-        static final String TOTAL_DOWNLOADS = "total_downloads";
+        static final String TOTAL_UPLOAD_SIZE = "total_upload_size";
+        static final String TOTAL_DOWNLOAD_SIZE = "total_download_size";
         static final String MAX_REFRESH_TIME_LAG = "max_refresh_time_lag";
         static final String MAX_REFRESH_TIME_LAG_IN_MILLIS = "max_refresh_time_lag_in_millis";
         static final String REFRESH_SIZE_LAG = "refresh_size_lag";